refactor(LottoGenerator): abstract class to interface
- lotto generator 추상화
- 패키지 서브 패키지로 이동
