+++ src/main/java/lotto/domain/lottogenerator/LottoGenerator.java
@@ -1,0 +1,8 @@
+package lotto.domain.lottogenerator;
+
+import java.util.Set;
+
+public interface LottoGenerator {
+
+    Set<Integer> generateLotto();
+}
